+++ src/Controller/OrderController.php
@@ -25,6 +25,7 @@ use OxidSolutionCatalysts\Unzer\Service\UnzerSDKLoader;
 use OxidSolutionCatalysts\Unzer\Traits\ServiceContainer;
 use OxidSolutionCatalysts\Unzer\Service\UnzerDefinitions;
 use OxidSolutionCatalysts\Unzer\Core\UnzerDefinitions as CoreUnzerDefinitions;
+use UnzerSDK\Exceptions\UnzerApiException;
 
 /**
  * TODO: Decrease count of dependencies to 13
@@ -62,6 +63,8 @@ class OrderController extends OrderController_parent
         // generate always a new threat metrix session id
         $unzer = $this->getServiceFromContainer(Unzer::class);
         $this->_aViewData['unzerThreatMetrixSessionID'] = $unzer->generateUnzerThreatMetrixIdInSession();
+        $this->_aViewData['uzrcurrency'] = $this->getActCurrency();
+        ;
         $this->getSavedPayment();
         return parent::render();
     }
@@ -244,7 +247,7 @@ class OrderController extends OrderController_parent
      */
     public function getPaymentSaveSetting()
     {
-        $bSavedPayment = Registry::getConfig()->getConfigParam('UnzerSavePaymentsForUser');
+        $bSavedPayment = 1;
 
         // no guests allowed
         $user = $this->getUser();
@@ -315,7 +318,7 @@ class OrderController extends OrderController_parent
         }
         return parent::getExecuteFnc();
     }
-    protected function getSavedPayment(): void
+    protected function getSavedPayment()
     {
         $UnzerSdk = $this->getServiceFromContainer(UnzerSDKLoader::class);
         $unzerSDK = $UnzerSdk->getUnzerSDK();
@@ -325,7 +328,11 @@ class OrderController extends OrderController_parent
         if ($ids) {
             foreach ($ids as $typeId) {
                 if (!empty($typeId['PAYMENTTYPEID'])) {
-                    $paymentType = $unzerSDK->fetchPaymentType($typeId['PAYMENTTYPEID']);
+                    try {
+                        $paymentType = $unzerSDK->fetchPaymentType($typeId['PAYMENTTYPEID']);
+                    } catch (UnzerApiException $e) {
+                        continue;
+                    }
 
                     if (strpos($typeId['PAYMENTTYPEID'], 'crd')) {
                         $paymentTypes['card'][$typeId['PAYMENTTYPEID']] = $paymentType->expose();