+++ src/Service/Payment.php
@@ -9,8 +9,10 @@ namespace OxidSolutionCatalysts\Unzer\Service;
 
 use Exception;
 use OxidEsales\Eshop\Application\Model\Order;
+use OxidEsales\Eshop\Application\Model\Basket as BasketModel;
 use OxidEsales\Eshop\Application\Model\Payment as PaymentModel;
 use OxidEsales\Eshop\Core\Registry;
+use OxidEsales\Eshop\Core\Request;
 use OxidEsales\Eshop\Core\Session;
 use OxidSolutionCatalysts\Unzer\Core\UnzerDefinitions;
 use OxidSolutionCatalysts\Unzer\Exception\Redirect;
@@ -19,10 +21,12 @@ use OxidSolutionCatalysts\Unzer\PaymentExtensions\UnzerPayment as AbstractUnzerP
 use OxidSolutionCatalysts\Unzer\Service\Transaction as TransactionService;
 use UnzerSDK\Exceptions\UnzerApiException;
 use UnzerSDK\Resources\AbstractUnzerResource;
+use UnzerSDK\Resources\Basket;
 use UnzerSDK\Resources\PaymentTypes\BasePaymentType;
 use UnzerSDK\Resources\PaymentTypes\InstallmentSecured;
 use UnzerSDK\Resources\TransactionTypes\Authorization;
 use UnzerSDK\Resources\TransactionTypes\Cancellation;
+use UnzerSDK\Resources\TransactionTypes\Charge;
 use UnzerSDK\Resources\TransactionTypes\Shipment;
 
 /**
@@ -96,6 +100,7 @@ class Payment
      */
     public function executeUnzerPayment(PaymentModel $paymentModel): bool
     {
+        $paymentExtension = null;
         try {
             /** @var string $customerType */
             $customerType = Registry::getRequest()->getRequestParameter('unzer_customer_type', '');
@@ -114,14 +119,6 @@ class Payment
                 $basket
             );
 
-            /** @var string $sess_challenge */
-            $sess_challenge = $this->session->getVariable('sess_challenge');
-            $this->transactionService->writeTransactionToDB(
-                $sess_challenge,
-                $this->session->getUser()->getId(),
-                $this->getSessionUnzerPayment()
-            );
-
             $paymentStatus = $this->getUnzerPaymentStatus() !== self::STATUS_ERROR;
 
             if ($this->redirectUrl) {
@@ -134,8 +131,6 @@ class Payment
         } catch (Redirect $e) {
             throw $e;
         } catch (UnzerApiException $e) {
-            $this->removeTemporaryOrder();
-
             throw new RedirectWithMessage(
                 $this->unzerService->prepareOrderRedirectUrl(
                     $paymentExtension instanceof AbstractUnzerPayment && $paymentExtension->redirectUrlNeedPending()
@@ -143,8 +138,6 @@ class Payment
                 $this->translator->translateCode($e->getErrorId(), $e->getClientMessage())
             );
         } catch (Exception $e) {
-            $this->removeTemporaryOrder();
-
             throw new RedirectWithMessage(
                 $this->unzerService->prepareOrderRedirectUrl(
                     $paymentExtension instanceof AbstractUnzerPayment && $paymentExtension->redirectUrlNeedPending()
@@ -170,10 +163,15 @@ class Payment
     /**
      * @return string
      * @throws UnzerApiException
+     * @SuppressWarnings(PHPMD.CyclomaticComplexity)
      */
     public function getUnzerPaymentStatus(): string
     {
         $result = self::STATUS_ERROR;
+        $sessionUnzerPayment = $this->getSessionUnzerPayment();
+        if (is_null($sessionUnzerPayment)) {
+            return $result;
+        }
 
         /** @var \UnzerSDK\Resources\Payment $sessionUnzerPayment */
         $sessionUnzerPayment = $this->getSessionUnzerPayment();
@@ -229,17 +227,19 @@ class Payment
      * @param string $customerType
      * @param string $currency
      * @return \UnzerSDK\Unzer
-     */
-    public function getUnzerSDK(string $customerType = '', string $currency = ''): \UnzerSDK\Unzer
+    */
+    protected function getUnzerSDK(string $customerType = '', string $currency = ''): \UnzerSDK\Unzer
     {
         return $this->unzerSDKLoader->getUnzerSDK($customerType, $currency);
     }
 
     /**
+     * @param null $paymentExtension
+     * @param null $currency
      * @return \UnzerSDK\Resources\Payment|null
      * @throws UnzerApiException
      */
-    public function getSessionUnzerPayment(): ?\UnzerSDK\Resources\Payment
+    public function getSessionUnzerPayment($paymentExtension = null, $currency = null): ?\UnzerSDK\Resources\Payment
     {
         $paymentId = $this->session->getVariable('PaymentId');
         if (is_string($paymentId)) {
@@ -248,19 +248,18 @@ class Payment
             /** @var Order $order */
             $order = oxNew(Order::class);
             $order->load($sessionOrderId);
-
-            $customerType = '';
+            $customerType = 'B2C';
             /** @var string $currency */
             $currency = $order->getFieldData('oxcurrency') ?? '';
-            if ($order->getFieldData('oxpaymenttype') == UnzerDefinitions::INVOICE_UNZER_PAYMENT_ID) {
-                $customerType = 'B2C';
-                if (!empty($order->getFieldData('oxbillcompany')) || !empty($order->getFieldData('oxdelcompany'))) {
-                    $customerType = 'B2B';
+
+            if ($paymentExtension !== null && $currency !== null) {
+                if ($order->getFieldData('oxpaymenttype') === UnzerDefinitions::INVOICE_UNZER_PAYMENT_ID) {
+                    if (Registry::getRequest()->getRequestParameter('unzer_customer_type') !== 'B2C') {
+                        $customerType = 'B2B';
+                    }
                 }
             }
-
-            $sdk = $this->unzerSDKLoader->getUnzerSDK($customerType, $currency);
-            return $sdk->fetchPayment($paymentId);
+            return $this->unzerSDKLoader->getUnzerSDK($customerType, $currency)->fetchPayment($paymentId);
         }
 
         return null;
@@ -334,8 +333,8 @@ class Payment
             );
             $payment = $charge->getPayment();
             if (
+                $charge->isSuccess() &&
                 ($charge->isSuccess()) &&
-                ($payment instanceof \UnzerSDK\Resources\Payment) &&
                 $payment->getAmount()->getRemaining() === 0.0
             ) {
                 $oOrder->markUnzerOrderAsPaid();